factor out FieldPropsByName type
