+++ protocol-designer/src/components/StepEditForm/forms/MoveLiquidForm/SourceDestHeaders.js
@@ -4,8 +4,8 @@ import { FormGroup } from '@opentrons/components'
 import { i18n } from '../../../../localization'
 import { LabwareField, WellSelectionField } from '../../fields'
 import { AspDispSection } from '../AspDispSection'
-import type { FieldPropsByName } from '../../fields/makeSingleEditFieldProps'
 import type { StepFieldName } from '../../../../steplist/fieldLevel'
+import type { FieldPropsByName } from '../../types'
 
 import styles from '../../StepEditForm.css'
