+++ protocol-designer/src/components/StepEditForm/forms/ThermocyclerForm/ProfileSettings.js
@@ -7,7 +7,7 @@ import { TextField } from '../../fields'
 
 import styles from '../../StepEditForm.css'
 
-import type { FieldPropsByName } from '../../fields/makeSingleEditFieldProps'
+import type { FieldPropsByName } from '../../types'
 
 type Props = {| propsForFields: FieldPropsByName |}
